# Use StringMemory instead of StringVector where possible